pyproject.toml
[tool.poetry]
name = "smart-api-monitor"
version = "0.1.0"
description = "Intelligent API Monitoring System with ML-powered analytics"
authors = ["Your Name <your.email@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.68.0"
uvicorn = "^0.15.0"
pandas = "^1.3.0"
numpy = "^1.21.0"
scikit-learn = "^0.24.2"
networkx = "^2.6.2"
pydantic = "^1.8.2"
python-dotenv = "^0.19.0"
PyYAML = "^5.4.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
black = "^21.7b0"
flake8 = "^3.9.2"
mypy = "^0.910"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"